refactor : UserService 인터페이스를 제거 및 구현체였던 클래스의 클래스명을 UserService로 변경
- 리뷰에서 왜 인터페이스로 구현했냐라는 질문을 받아서 생각해 본 결과, 서비스가 변경될거라는 요구사항이 없어서 굳이 인터페이스로 구현할 필요가 없다는걸 알게되어 변경했습니다.
