+++ src/main/java/com/kakao/cafe/service/user/UserService.java
@@ -2,17 +2,46 @@ package com.kakao.cafe.service.user;
 
 import com.kakao.cafe.domain.user.User;
 import com.kakao.cafe.domain.user.UserDto;
+import com.kakao.cafe.repository.user.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
-public interface UserService {
+@Service
+public class UserService {
 
-    void createUser(UserDto userDto);
+    private UserRepository repository;
 
-    Optional<User> findSingleUser(String userId);
+    @Autowired
+    public UserService(UserRepository repository) {
+        this.repository = repository;
+    }
+    
+    public void createUser(UserDto userDto) {
+        String userId = userDto.getUserId();
+        String password = userDto.getPassword();
+        String email = userDto.getEmail();
+        String name = userDto.getName();
 
-    List<User> findAllUsers();
+        User user = new User(userId, password, email, name);
 
-    void isDuplicatedUser(User user);
+        isDuplicatedUser(user);
+        repository.save(user);
+    }
+
+    public Optional<User> findSingleUser(String userId) {
+        return repository.findByUserId(userId);
+    }
+
+    public List<User> findAllUsers() {
+        return repository.findAll();
+    }
+
+    public void isDuplicatedUser(User user) {
+        if (repository.findByUserId(user.getUserId()).isPresent()) {
+            throw new IllegalStateException("사용자 ID가 중복됩니다.");
+        }
+    }
 }
